Added converter, to convert from v7 to v8. All v7 code is marked deprecated.
To convert, type: /tchest convert

This is called version 8.0 beta.
It will be 8.0 when the converter is removed.

+++ src/main/java/com/mtihc/minecraft/treasurechest/v8/core/TreasureChest.java
@@ -68,7 +68,7 @@ public class TreasureChest implements ITreasureChest {
 		
 		unlimited = (Boolean) values.get("unlimited");
 		random = (Integer) values.get("random");
-		forgetTime = (long) (Integer) values.get("forget-time");
+		forgetTime = (Long) values.get("forget-time");
 		ignoreProtection = (Boolean) values.get("ignore-protection");
 		
 		Map<?, ?> rewardSection = (Map<?, ?>) values.get("rewards");
